==> 03-cma.Rmd <==
---
title: "CMA"
author: "Paloma Cárcamo"
output: html_document
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
if (!require("pacman")) install.packages("pacman")
pacman::p_load(tidyverse, CMAverse, gtsummary)
```

## Load data

```{r}
data_raw <- read_rds("data/joined-db-clean.rds")

data <- data_raw |> 
  mutate(parity = cut(children, 
                      breaks = c(0, 1, 3, 100), 
                      labels = c("0-1", "2-3", "3+"), 
                      include.lowest = TRUE),
         inc_cat = cut(income, 
                       breaks = quantile(data_raw$income, c(0, 0.25, 0.5, 0.75, 1), na.rm = TRUE), 
                       labels = c("Q1", "Q2", "Q3", "Q4"),
                       include.lowest = TRUE),
         bmi_cat = cut(bmi,
                       breaks = c(0, 25, 30, 1000),
                       labels = c("Normal or underweight", "Overweight", "Obesity")),
         smoking = factor(smoking),
         mi = factor(mi),
         oc_use = factor(oc_use),
         hbp = factor(hbp),
         hbp_med = factor(hbp_med),
         cohort = factor(cohort)) |> 
  select(mi, oc_use, hbp, cohort, smoking, inc_cat, parity, bmi_cat, hbp_med) |> 
  na.omit()

data_withnas <- data_raw |> 
  mutate(parity = cut(children, 
                      breaks = c(0, 1, 3, 100), 
                      labels = c("0-1", "2-3", "3+"), 
                      include.lowest = TRUE),
         inc_cat = cut(income, 
                       breaks = quantile(data_raw$income, c(0, 0.25, 0.5, 0.75, 1), na.rm = TRUE), 
                       labels = c("Q1", "Q2", "Q3", "Q4"),
                       include.lowest = TRUE),
         bmi_cat = cut(bmi,
                       breaks = c(0, 25, 30, 1000),
                       labels = c("Normal or underweight", "Overweight", "Obesity")),
         smoking = factor(smoking),
         mi = factor(mi),
         oc_use = factor(oc_use),
         hbp = factor(hbp),
         hbp_med = factor(hbp_med),
         cohort = factor(cohort),
         length_ocuse = oc_yearstop - oc_yearstart) |> 
  select(mi, oc_use, hbp, cohort, smoking, inc_cat, parity, bmi_cat, hbp_med, length_ocuse)

length_ocuse <- data_raw |> 
  mutate(parity = cut(children, 
                      breaks = c(0, 1, 3, 100), 
                      labels = c("0-1", "2-3", "3+"), 
                      include.lowest = TRUE),
         inc_cat = cut(income, 
                       breaks = quantile(data_raw$income, c(0, 0.25, 0.5, 0.75, 1), na.rm = TRUE), 
                       labels = c("Q1", "Q2", "Q3", "Q4"),
                       include.lowest = TRUE),
         bmi_cat = cut(bmi,
                       breaks = c(0, 25, 30, 1000),
                       labels = c("Normal or underweight", "Overweight", "Obesity")),
         smoking = factor(smoking),
         mi = factor(mi),
         oc_use = factor(oc_use),
         hbp = factor(hbp),
         hbp_med = factor(hbp_med),
         cohort = factor(cohort)) |> 
  select(id, cohort, mi, oc_use, hbp, cohort, smoking, inc_cat, parity, bmi_cat, hbp_med) |> 
  na.omit() |> 
  left_join(data_raw |> select(id, cohort, oc_yearstart, oc_yearstop), by = c("id", "cohort")) |> 
  mutate(length_ocuse = oc_yearstop - oc_yearstart) |> 
  select(-id, -cohort, -oc_yearstart, -oc_yearstop)
```

```{r}
data2_raw <- read_rds("data/joined-db2-clean.rds")

data2 <- data2_raw |> 
  mutate(parity = cut(children, 
                      breaks = c(0, 1, 3, 100), 
                      labels = c("0-1", "2-3", "3+"), 
                      include.lowest = TRUE),
         inc_cat = cut(income, 
                       breaks = quantile(data_raw$income, c(0, 0.25, 0.5, 0.75, 1), na.rm = TRUE), 
                       labels = c("Q1", "Q2", "Q3", "Q4"),
                       include.lowest = TRUE),
         bmi_cat = cut(bmi,
                       breaks = c(0, 25, 30, 1000),
                       labels = c("Normal or underweight", "Overweight", "Obesity")),
         smoking = factor(smoking),
         mi = factor(mi),
         oc_use = factor(oc_use),
         hbp = factor(hbp),
         hbp_med = factor(hbp_med),
         cohort = factor(cohort)) |> 
  select(mi, oc_use, hbp, cohort, smoking, inc_cat, parity, bmi_cat, hbp_med) |> 
  na.omit()
```

## Descriptive table

```{r}
## Only complete observations
tbl_summary(data = data, by = mi) |> 
  add_p()

## Only complete observations, with length of OC use
tbl_summary(data = length_ocuse, by = mi) |> 
  add_p()

## All observations
tbl_summary(data = data_withnas, by = mi) |> 
  add_p()
```

```{r}
tbl_summary(data = data2, by = mi) |> 
  add_p()

xtabs(~mi + hbp + cohort, data = data2)
```

## Outcome regression

```{r}
mod1 <- glm(mi ~ oc_use + cohort + smoking + inc_cat + parity, family = binomial, data = data)

summary(mod1)
```

## Causal mediation analysis

```{r}
set.seed(11112)

res_msm <- cmest(data = data,
                 model = "msm", 
                 outcome = "mi", 
                 exposure = "oc_use",
                 mediator = "hbp", 
                 basec = c("cohort", "smoking", "inc_cat", "parity", "bmi_cat", "hbp_med"), 
                 EMint = TRUE,
                 
                 ereg =  glm(oc_use ~ cohort + smoking + inc_cat + parity, family = binomial, data = data), # Exposure regression for weighting (IPW for exposure)
                 
                 wmnomreg = list(glm(hbp ~ 1, family = binomial, data = data)), # IPW for mediator (numerator)
                 
                 wmdenomreg = list(glm(hbp ~ oc_use + cohort + smoking + inc_cat + parity + bmi_cat, family = binomial, data = data)), # IPW for mediator (denominator)
                 
                 yreg = glm(mi ~ oc_use + hbp + oc_use*hbp, family = binomial, data = data),   # Outcome regression for CDE
                 
                 mreg = list(glm(hbp ~ oc_use, family = binomial, data = data)), # Mediator regression
                 
                 astar = 0, 
                 a = 1, 
                 mval = list(0), 
                 estimation = "imputation", 
                 inference = "bootstrap", 
                 nboot = 10000)

# save(res_msm, file = "data/result-msm.RData")
summary(res_msm)
```

## IPMW

```{r}
summary(glm(mi ~ hbp, data = data_withnas, family = binomial))
summary(glm(mi ~ hbp, data = data, family = binomial))
```
